requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
pydantic>=2.6.1
pydantic-settings>=2.2.1
httpx>=0.27.0
python-binance>=1.0.0
python-dateutil>=2.8.2
tenacity>=8.2.3
loguru>=0.7.2
redis>=5.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

